apps/ai-engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
groq>=0.4.0
httpx>=0.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
tenacity>=8.2.0
asyncpg>=0.29.0
